requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
gcloud==0.17.0
googleapis-common-protos==1.63.0
httplib2==0.22.0
idna==3.7
jwcrypto==1.5.6
jws==0.1.3
oauth2client==3.0.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.4.3
pyparsing==3.1.2
python-jwt==2.0.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.9
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==1.26.18
